feat: add protected route component
