+++ src/components/Admin/Admin.tsx
@@ -1,15 +1,10 @@
 import {useTranslation} from "react-i18next";
-import React, {useEffect} from "react";
+import React from "react";
 import "../../styles/Admin.css";
 import ButtonsJavaEdition from "../utilities/ButtonsJavaEdition";
-import {useNavigate} from "react-router-dom";
-import {forbidAccessToAdminResource} from "../../utils";
 
 function Admin() {
-    const navigate = useNavigate();
     const {t} = useTranslation();
-
-    useEffect(() => {forbidAccessToAdminResource(navigate)}, [navigate]);
 
     return (
         <div className="admin-page">
